tests, docs, and datatypes, oh my!
Added remainder of reference chacha tests (WE PASS ALL OF THEM LFGGGGGGGGGGGGGG), text_* tests now test odd length strings as well, doc improvements, and now use unsigned data in `Row`. This makes no difference to any code but tests, which need to set values that would overflow if the values were signed (and this pisses rustc off), so unsigned it is.