AVX512 support, gated behind `nightly` flag
Just couldn't stop myself. Not sure if GitHub runners will be able to run the AVX512 test, but I was able to verify this passes all tests on my machine using Intel's SDE for emulating AVX512 support.

Unfortunately, this approach makes it impossible to run the comparisons, so I can't get a general sense of how much faster AVX512 is over AVX2.  I suspect it's more than the SSE2->AVX2 jump, since we no longer need to emulate rotate instructions (finally), and the whole set of 4 blocks is computed in a single pass instead of requiring a loop.